Add avatar dropdown menu and some tables
This commit adds a dropdown menu to the avatar button conataining
the "quick links" of the webpage. This quick links will only be
accessible through this dropdown menu if the width of the screen
is too small for all the options to be visible on the header bar,
and will disappear from view.

The dropdown menu can only be accessed through mouse hover. Accessing
through onClick of the avatar button might require JavaScript.

The commit also adds some tables listing the blogs and the talks
which also have stylized buttons to read or view them respectively.

Corrects the focus order of the elements of the webpage. The page
will now allow focusing of headers of each section.

Also adds a slight margin on the bottom of the body above the footer
to provide enough spacing for text (if any)at the bottom to be visible.