Add global styling variables, some responsiveness and change table to divs.
This commit adds some global styling variables ranging from the primary colors
to the font and some element sizes which is used throughout the stylesheet.

The table used in the page has been changed to divs for better responsiveness.

The dropdown menu now "correctly" appears on top of the footer. Also adds
scrolling ability in case the height is too small to display all the options.

Fix the padding and sizes of the content with respect to the footer.