Add projects, change intro and fix flex basis for some elements
This commit adds all the projects and changes the links which
stopped working or are replaced by links that would be more
appropriate.

Also updates the intro section, removing the dummy text.

The flex-basis of some elements(for instance, the buttons on main
content) with `flex-basis: 0px` have been changed to
`flex-basis: auto` for better responsiveness.